pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Daniel Ferreira <danielferreiranunes2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12, <4.0"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0"
python-dotenv = "^1.0"
pydantic = "^2.0"
pydantic-settings = "^2.0"
alembic = "^1.12.0"  
psycopg2-binary = "^2.9"
email-validator = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
httpx = "^0.25.0"